Optimize Homekit response time

+++ main/CoreBridge/CConnector.cpp
@@ -5,16 +5,13 @@
 extern "C"
 {
 #endif
-  void CoreBridge_doModulesAction(uint8_t *addrs, uint8_t *actions, uint8_t length)
-  {
+  void CoreBridge_doModulesAction(uint8_t* addrs, uint8_t* actions, uint8_t length)   {
     CoreBridge.doModulesAction(addrs, actions, length, true);
   }
 
-  int CoreBridge_getModuleAddrByHc(hap_char_t *hc)
-  {
-    for (int i = 0; i < CoreBridge.getModuleNum(); i++)
-    {
-      if (hap_char_get_iid((hap_char_t *)CoreBridge.getModule(i)->hc) == hap_char_get_iid(hc))
+  uint8_t CoreBridge_getModuleAddrByHc(hap_char_t* hc)   {
+    for (int i = 0; i < CoreBridge.getModuleNum(); i++)     {
+      if (hap_char_get_iid((hap_char_t*)CoreBridge.getModule(i)->hc) == hap_char_get_iid(hc))
         return i + 1;
     }
